requirements.txt
google-genai>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
langgraph>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
tavily-python>=0.5.0
duckduckgo-search>=6.0.0
